Refactor duplicate code to a utility function
